Moved CCoins, CTxOutCompressor, CTxInUndo, and CTxUndo to core. Moved UpdateTime out of CBlockHeader and moved CBlockHeader into core.
